moved save as dialog logic into separated file
